fix: update IJobExecutionRepository interface to use int instead of long
Fixed compilation error CS0535 where JobExecutionRepository did not implement
the interface member IJobExecutionRepository.GetByIdAsync(long).

Changed GetByIdAsync(long id) to GetByIdAsync(int id) in the interface to match:
- The repository implementation (changed in previous commit)
- The database schema (JobExecutions.Id is INT, not BIGINT)
- The entity model (JobExecution now inherits int Id from BaseEntity)

This completes the fix for the InvalidCastException issue.

+++ src/SchedulerPlatform.Core/Domain/Interfaces/IJobExecutionRepository.cs
@@ -5,7 +5,7 @@ namespace SchedulerPlatform.Core.Domain.Interfaces;
 
 public interface IJobExecutionRepository : IRepository<JobExecution>
 {
-    Task<JobExecution?> GetByIdAsync(long id);
+    Task<JobExecution?> GetByIdAsync(int id);
     Task<IEnumerable<JobExecution>> GetByScheduleIdAsync(int scheduleId);
     Task<IEnumerable<JobExecution>> GetByStatusAsync(JobStatus status);
     Task<IEnumerable<JobExecution>> GetFailedExecutionsAsync(int scheduleId);
